Add --pull option to podman run and create
Upstream docker has a --pull option for run and create, which updates an image in storage if it is out of date.
Add this option to podman, by reworking some of libpod/image/pull.go to allow for a source image to be inspected
prior to being pulled, and checked if the pull needs to happen.

Signed-off-by: Peter Hunt <pehunt@redhat.com>