podman umount: error out if called with no args
Inspired by #2684, I wrote a CI test to look for other such
instances in which a command is invoked without a required
argument. 'podman umount' seems to be the only one, and
solution is simple: checkAllAndLatest() already does the
check for us.

Resolve a few other problems uncovered by testing:

  podman mount: indicate that CONTAINER arg is optional

  podman pod stats: ditto

  podman generate kube: remove check for -l (latest) flag,
    it isn't actually implemented.

Signed-off-by: Ed Santiago <santiago@redhat.com>