docs/podman.1.md: rename "containers-registries.conf" to "registries.conf"
Since the correct name for the registries man page is the shorter
"registries.conf", correct the reference in the "SEE ALSO" section.

Note that this is a temporary solution since, in order to avoid possible
future clashes with more generic man pages, all of the podman-related
man pages should ideally start with a "containers-" prefix.

Signed-off-by: Robert P. J. Day <rpjday@crashcourse.ca>