Re-allow the : character in container names, but warn
Per discussion in #2876, our initial container name regular
expression allowed the : character (more specifically, it just
didn't force the match to extend to the entirety of the container
name, so any name containing at least one alphanumeric character
was considered valid). The Fedora Toolbox used names including
the : character, so making the regex actually match the entire
container name broke them.

Make : an allowed character again, but warn that it is deprecated
and may be removed.

Fixes: #2876

Signed-off-by: Matthew Heon <matthew.heon@pm.me>