oci: fix race when the container exits before start returns
fix a race when the container process exits faster than we are
notified from the OCI runtime start command.  When it happens, the
cleanup process exits immediately as the container is reported as
running.  Do the check while we hold the lock, so we won't release it
with an invalid status.

Reproducer:
NOTIFY_SOCKET=/run/user/1000/notify podman run fedora sh -c ' echo hello'

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>