Cirrus: Temp. workaround VFS incompatible option
There are many integration tests using the VFS storage driver.  A
recent update to the containers-common package in Fedora added a new
mount option `metacopy=on` as it enhances security.  However, this option
is not compatible with the VFS storage driver, causing those tests to
fail.

Until a better solution is realized, workaround this problem by forcing
"nodev" as the only mount option on all platforms.  Also add a visible
note to the output flagging the change.

Signed-off-by: Chris Evich <cevich@redhat.com>