Don't let Systemd kill Conmon on shutdown
If we set SIGUSR1, we won't kill Conmon or force a SIGTERM to be
sent to the container. Instead, the container will exit as per
usual during system shutdown, and conmon will remain alive to
record its exit status.

Use a 10 minute timeout so we don't permanently halt system
shutdown if an error occurs.

Signed-off-by: Matthew Heon <matthew.heon@pm.me>